+++ Sources/gestion-culturelle-back/src/main/java/com/afpa/cda/service/impl/SalleServiceImpl.java
@@ -29,16 +29,16 @@ public class SalleServiceImpl implements ISalleService {
 
 	@Autowired
 	private SalleRepository salleRepository;
-	
+
 	@Autowired
 	private TypeSalleRepository 	typeSalleRepository;
-	
+
 	@Autowired
 	private AnimationRepository animationRepository;
 
 	@Autowired
 	private ManifestationRepository manifestationRepository;
-	
+
 	@Autowired
 	private ModelMapper modelMapper;
 
@@ -58,73 +58,83 @@ public class SalleServiceImpl implements ISalleService {
 								.label(s.getTypesalle().getLabel())
 								.build())
 						.build())
-						.collect(Collectors.toList());
+				.collect(Collectors.toList());
 	}
-
 
 	@Override
-	public SalleDto add(SalleDto salleDto) {
-		
-		Salle salle = new Salle ();
-		salle.setLabel(salleDto.getLabel());
-		salle.setCapacite(salleDto.getCapacite());
-		salle.setPlacesVip(salleDto.getPlacesVip());
-		salle.setFraisjournalier(salleDto.getFraisJournalier());
-		
-		Optional <TypeSalle> typesalleOp = this.typeSalleRepository.findById(salleDto.getTypeSalle().getId());
-		if (typesalleOp.isPresent()) {
-			TypeSalle typesalle = typesalleOp.get();
-			TypeSalleDto typesalleDto = modelMapper.map(typesalle,TypeSalleDto.class);			
-			
-			salleDto.setTypeSalle(typesalleDto);
+	public boolean add(SalleDto salleDto) {
+		List <Salle> listSalles = this.salleRepository.findAll();
+		boolean salleExistant = false;
+		for (Salle salle : listSalles) {
+			if (salle.getLabel().equalsIgnoreCase(salleDto.getLabel())) {
+				salleExistant = true;
+			}
 		}
-		
-		List<ManifestationDto> listManifestations = new ArrayList<ManifestationDto>() ;
-		
-		salleDto.setManifestations(listManifestations);
-		
-		this.salleRepository.save(this.modelMapper.map(salleDto,Salle.class));
-		
-		
-		return salleDto;
-		
-		
-//		try {
-//		Salle salle = this.salleRepository.save(this.modelMapper.map(salleDto,Salle.class));
-//		salleDto.setId(salle.getId());
-//		System.err.println("Salle ajoutee");
-//		} catch (Exception e) {
-//			System.err.println(e.getStackTrace());
-//		}
-//		return salleDto;
-//		
+
+		if (!salleExistant) {
+
+			Salle salle = new Salle ();
+			salle.setLabel(salleDto.getLabel());
+			salle.setCapacite(salleDto.getCapacite());
+			salle.setPlacesVip(salleDto.getPlacesVip());
+			salle.setFraisjournalier(salleDto.getFraisJournalier());
+
+			Optional <TypeSalle> typesalleOp = this.typeSalleRepository.findById(salleDto.getTypeSalle().getId());
+			if (typesalleOp.isPresent()) {
+				TypeSalle typesalle = typesalleOp.get();
+				TypeSalleDto typesalleDto = modelMapper.map(typesalle,TypeSalleDto.class);			
+
+				salleDto.setTypeSalle(typesalleDto);
+			}
+
+			//		List<ManifestationDto> listManifestations = new ArrayList<ManifestationDto>() ;
+
+			//		salleDto.setManifestations(listManifestations);
+
+			this.salleRepository.save(this.modelMapper.map(salleDto,Salle.class));
+
+
+			return salleExistant;
+
+
+			//		try {
+			//		Salle salle = this.salleRepository.save(this.modelMapper.map(salleDto,Salle.class));
+			//		salleDto.setId(salle.getId());
+			//		System.err.println("Salle ajoutee");
+			//		} catch (Exception e) {
+			//			System.err.println(e.getStackTrace());
+			//		}
+			//		return salleDto;
+			//		
+		}
+		return salleExistant;
 	}
 
-	
+
 	@Override
 	public boolean updateSalle(SalleDto salleDto, int id) {
 		Optional <Salle> salleOp= this.salleRepository.findById(id);
 		if(salleOp.isPresent()) {
 			Salle salle = salleOp.get();
 			salle=this.modelMapper.map(salleDto,Salle.class);
-		
-			List<Manifestation> listManifestation = manifestationRepository.findAll();
-			for (Manifestation manifestation : listManifestation) {
-				if (manifestation.getSalle().getId()==salle.getId()) {
-					ManifestationDto manifestationDto = modelMapper.map(manifestation,ManifestationDto.class);
-					manifestationDto=calcul(manifestationDto);
-					Manifestation manif = modelMapper.map(manifestationDto,Manifestation.class);
-					
-					Optional<Manifestation> manifOp = this.manifestationRepository.findById(manifestation.getId());
-					if (manifOp.isPresent()) {
-						this.manifestationRepository.save(manif);
-					}
-					
-				}
-			}
-			
+
+			//			List<Manifestation> listManifestation = manifestationRepository.findAll();
+			//			for (Manifestation manifestation : listManifestation) {
+			//				if (manifestation.getSalle().getId()==salle.getId()) {
+			//					ManifestationDto manifestationDto = modelMapper.map(manifestation,ManifestationDto.class);
+			//					manifestationDto=calcul(manifestationDto);
+			//					Manifestation manif = modelMapper.map(manifestationDto,Manifestation.class);
+			//					
+			//					Optional<Manifestation> manifOp = this.manifestationRepository.findById(manifestation.getId());
+			//					if (manifOp.isPresent()) {
+			//						this.manifestationRepository.save(manif);
+			//					}
+			//					
+			//				}
+			//			}
+
 			this.salleRepository.save(salle);
-			
+
 			System.err.println("Salle mise à jour");
 			return true;
 		}
@@ -133,7 +143,7 @@ public class SalleServiceImpl implements ISalleService {
 
 	@Override
 	public ManifestationDto calcul (ManifestationDto manifDto) {
-		
+
 		double debut=manifDto.getDateDebut().getTime();
 		double fin=manifDto.getDateFin().getTime();
 		double duree = 1+((((fin-debut)/1000)/3600)/24);
@@ -153,21 +163,33 @@ public class SalleServiceImpl implements ISalleService {
 
 		manifDto.setReservations(animDto.getNbreSpectateursPrevus());
 		manifDto.setReservationsVip(salleDto.getPlacesVip());
-		
-	
+
+
 		manifDto.setCout( (animDto.getPrix()+(duree* salleDto.getFraisJournalier())));
 		manifDto.setPrixBillet(manifDto.getCout()/(animDto.getNbreSpectateursPrevus()*0.8));
-		
+
 		return manifDto;
 	}
-	
+
 	@Override
 	public boolean deleteSalle(int id) {
-		if(this.salleRepository.existsById(id)) {
+		List <Manifestation> listManifestations = manifestationRepository.findAll(); 
+		boolean manifestationAvecSalle = false;
+		for (Manifestation manifestation : listManifestations) {
+			if (manifestation.getSalle().getId()==id) {
+				manifestationAvecSalle = true;
+				System.err.println(manifestationAvecSalle);
+				System.err.println("Salle non supprimee");
+			}
+		}
+
+		if(this.salleRepository.existsById(id) && !manifestationAvecSalle) {
 			this.salleRepository.deleteById(id);
 			System.err.println("Salle supprimee");
+			System.err.println(manifestationAvecSalle);
 			return true;
 		}
+		System.err.println(manifestationAvecSalle);
 		return false;
 	}
 
@@ -177,18 +199,18 @@ public class SalleServiceImpl implements ISalleService {
 		SalleDto salleDto = new SalleDto();
 		if(salleOp.isPresent()) {
 			Salle salle = salleOp.get();
-		
+
 			salleDto.setId(salle.getId());
 			salleDto.setLabel(salle.getLabel());
 			salleDto.setCapacite(salle.getCapacite());
 			salleDto.setPlacesVip(salle.getPlacesVip());
 			salleDto.setFraisJournalier(salle.getFraisjournalier());
-		
+
 			TypeSalleDto typeSalleDto = new TypeSalleDto();
 			typeSalleDto.setId(salle.getTypesalle().getId());
 			typeSalleDto.setLabel(salle.getTypesalle().getLabel());
 			salleDto.setTypeSalle(typeSalleDto);
-			
+
 		}
 		return salleDto;
 	}
